+++ ui/app/providers.tsx
@@ -5,7 +5,11 @@ import { useRouter } from "next/navigation";
 import { SessionProvider } from "next-auth/react";
 import { ThemeProvider as NextThemesProvider } from "next-themes";
 import { ThemeProviderProps } from "next-themes/dist/types";
+import posthog from "posthog-js";
+import { PostHogProvider as PHProvider } from "posthog-js/react";
 import * as React from "react";
+
+import { initializePostHog } from "@/lib/analytics";
 
 export interface ProvidersProps {
   children: React.ReactNode;
@@ -15,10 +19,17 @@ export interface ProvidersProps {
 export function Providers({ children, themeProps }: ProvidersProps) {
   const router = useRouter();
 
+  // Initialize PostHog
+  React.useMemo(() => {
+    initializePostHog();
+  }, []);
+
   return (
     <SessionProvider>
       <NextUIProvider navigate={router.push}>
-        <NextThemesProvider {...themeProps}>{children}</NextThemesProvider>
+        <NextThemesProvider {...themeProps}>
+          <PHProvider client={posthog}>{children}</PHProvider>
+        </NextThemesProvider>
       </NextUIProvider>
     </SessionProvider>
   );